Trekk ut kode i egne komponenter: Delvilkår.tsx og Unntak.tsx
